More memory opts

+++ pipeline/submodules/select_direction.py
@@ -70,10 +70,8 @@ def refusal_score(
     refusal_toks: Int[Tensor, 'batch seq'],
     epsilon: Float = 1e-8,
 ):
-    logits = logits.to(torch.float64)
-
-    # we only care about the last tok position
-    logits = logits[:, -1, :]
+    # slice to last position before casting to avoid OOM on large vocab/long sequences
+    logits = logits[:, -1, :].to(torch.float64)
 
     probs = torch.nn.functional.softmax(logits, dim=-1)
     refusal_probs = probs[:, refusal_toks].sum(dim=-1)
